+++ test/audits/metatags.test.js
@@ -273,6 +273,9 @@ describe('Meta Tags', () => {
         site,
         finalUrl: 'http://example.com',
         audit,
+        opportunity: {
+          setUpdatedBy: sinon.stub(),
+        },
       };
     });
 
@@ -428,6 +431,7 @@ describe('Meta Tags', () => {
           getType: () => 'meta-tags',
           setData: () => {},
           getData: () => {},
+          setUpdatedBy: sinon.stub().returnsThis(),
         };
         logStub = {
           info: sinon.stub(),
@@ -569,6 +573,7 @@ describe('Meta Tags', () => {
           remove: sinon.stub(),
           setData: sinon.stub(),
           save: sinon.stub(),
+          setUpdatedBy: sinon.stub().returnsThis(),
         };
 
         opportunity.getSuggestions.returns([existingSuggestion]);
@@ -698,6 +703,7 @@ describe('Meta Tags', () => {
           getType: () => 'meta-tags',
           setData: sinon.stub(),
           getData: sinon.stub(),
+          setUpdatedBy: sinon.stub().returnsThis(),
         };
 
         site = {